requirements.txt
tweepy
asyncio
python-dotenv
pytest
google-generativeai
